fix(domain): apply the Codex review of #15
- GitActivity::medianLinesChanged is now ?float, so bohort's fractional
  median (251.5) can be represented without a TypeError or precision loss.
- LevelRuleResult::isConfirmed() is false as soon as any verdict carries a
  Range, even when a lower Confirmed axis masks it and the global floor
  and ceiling collapse to the same level.
- ProfileIdentity carries the optional profile.json note (e.g. why
  declaratif.md is absent), so the renderer can cite it later.
- ProfileNotAssessable carries the partial ProfileIdentity when it was
  already read before the gate broke, so a NotAssessable Assessment can
  report what could be read anyway.
- Profile keeps the actual directory-piece inventory (presentPieces),
  alongside ProfileIdentity::available, so the declared/present mismatch
  can be reported for pieces this aggregate does not model (code/,
  session.md).
- SizeThresholds gains a SizeBand enum and bandForFiles()/bandForLines();
  InterventionThresholds and ParallelismThresholds gain levelForMedian().
  New tests pin every band/threshold boundary named in the review
  (files 2/3/8/9/20/21, lines 60/61/210/211/1000/1001, intervention
  3/2/1/0, parallelism 0/1/2/3, sample floors 5/12/5), plus a LevelRule
  regression test for the masked-Range case.

+++ src/Domain/Profile/GitActivity.php
@@ -16,7 +16,7 @@ final readonly class GitActivity
         public ?Period $period,
         public ?int $pullRequestsTotal,
         public ?int $medianFilesChanged,
-        public ?int $medianLinesChanged,
+        public ?float $medianLinesChanged,
         public ?int $medianCorrectionCommitsAfterOpen,
         public ?int $mergedWithoutHumanEditAfterOpen,
         public ?float $aiCoauthoredRatio,
